[web] Improve faste delete experience
 * Fix the exceptions thrown on attempting faste delete in empty strings
 * ComposeCommandCommunicator is TextLayoutResult-agnostic
